Refactor filesystem and serialization ops
This refactors all filesystem and serialization operations into a
separate LogFile struct to simplify testing. This would, were it a more
serious project, allow me to test the KV storage layer with various
injected FS/encoding faults (were I to provide a proper builder for the
KvStore), and test the FS accesses as a part of the LogFile on its own.

It also clarifies the implementation greatly: the KvStore layer is
responsible for mapping keys to LogFiles and maintaining indices whereas
the LogFile is responsible for read and write operations to disk.