Implement compaction rollback to handle faults
If an error is encountered while compacting the log directory, the
KvStore will be rolled back to the epoch it started at.

All log files for epochs later than the epoch being rolled-back to will
be delete from disc to prevent later stores from entering an
inconsistent state.

As there is no way to my knowledge to do a transactional delete in
linux, there's not much I can do if we fail to delete one of the files.

We could, potentially, have all log files created during compaction be
store in another directory then atomically move them into the current
one only after compaction is complete, but there's still no good way to
do a batch move then batch delete to unsure that we remain in a safe state